商品列表分页查询

+++ spring-cloud-woniu-manage/category-service/src/main/java/com/wcq/categoryservice/dao/mysql/po/CategoryPo.java
@@ -23,11 +23,11 @@ public class CategoryPo implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private BigInteger id;
+    private Long id;
 
     private String name;
 
-    private BigInteger parentId;
+    private Long parentId;
     private Integer isParent;
     private Integer sort;
 }
